backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.1
favicon==0.7.1
